Refactoring: extract generator for functions from `Any`
