+++ compiler/backend-common/src/org/jetbrains/kotlin/backend/common/DataClassMethodGenerator.kt
@@ -33,31 +33,21 @@ import org.jetbrains.kotlin.resolve.BindingContextUtils
  * TODO: data class with zero components gets no toString/equals/hashCode methods. This is inconsistent and should be
  * changed here with the platform backends adopted.
  */
-abstract class DataClassMethodGenerator(protected val declaration: KtClassOrObject, private val bindingContext: BindingContext) {
-    protected val classDescriptor: ClassDescriptor = BindingContextUtils.getNotNull(bindingContext, BindingContext.CLASS, declaration)
+abstract class DataClassMethodGenerator(
+    declaration: KtClassOrObject,
+    bindingContext: BindingContext
+) : FunctionsFromAnyGenerator(declaration, bindingContext) {
 
-    fun generate() {
+    override fun generate() {
         generateComponentFunctionsForDataClasses()
-
         generateCopyFunctionForDataClasses(primaryConstructorParameters)
 
-        val properties = dataProperties
-        if (properties.isNotEmpty()) {
-            generateDataClassToStringIfNeeded(properties)
-            generateDataClassHashCodeIfNeeded(properties)
-            generateDataClassEqualsIfNeeded(properties)
-        }
+        super.generate()
     }
 
     protected abstract fun generateComponentFunction(function: FunctionDescriptor, parameter: ValueParameterDescriptor)
 
     protected abstract fun generateCopyFunction(function: FunctionDescriptor, constructorParameters: List<KtParameter>)
-
-    protected abstract fun generateToStringMethod(function: FunctionDescriptor, properties: List<PropertyDescriptor>)
-
-    protected abstract fun generateHashCodeMethod(function: FunctionDescriptor, properties: List<PropertyDescriptor>)
-
-    protected abstract fun generateEqualsMethod(function: FunctionDescriptor, properties: List<PropertyDescriptor>)
 
     private fun generateComponentFunctionsForDataClasses() {
         // primary constructor should exist for data classes
@@ -76,38 +66,4 @@ abstract class DataClassMethodGenerator(protected val declaration: KtClassOrObje
         val copyFunction = bindingContext.get(BindingContext.DATA_CLASS_COPY_FUNCTION, classDescriptor) ?: return
         generateCopyFunction(copyFunction, constructorParameters)
     }
-
-    private fun generateDataClassToStringIfNeeded(properties: List<PropertyDescriptor>) {
-        val function = getMemberToGenerate(
-            classDescriptor, "toString",
-            KotlinBuiltIns::isString, List<ValueParameterDescriptor>::isEmpty
-        ) ?: return
-        generateToStringMethod(function, properties)
-    }
-
-    private fun generateDataClassHashCodeIfNeeded(properties: List<PropertyDescriptor>) {
-        val function = getMemberToGenerate(
-            classDescriptor, "hashCode",
-            KotlinBuiltIns::isInt, List<ValueParameterDescriptor>::isEmpty
-        ) ?: return
-        generateHashCodeMethod(function, properties)
-    }
-
-    private fun generateDataClassEqualsIfNeeded(properties: List<PropertyDescriptor>) {
-        val function = getMemberToGenerate(
-            classDescriptor, "equals",
-            KotlinBuiltIns::isBoolean
-        ) { parameters ->
-            parameters.size == 1 && KotlinBuiltIns.isNullableAny(parameters.first().type)
-        } ?: return
-        generateEqualsMethod(function, properties)
-    }
-
-    private val dataProperties: List<PropertyDescriptor>
-        get() = primaryConstructorParameters
-            .filter { it.hasValOrVar() }
-            .map { bindingContext.get(BindingContext.PRIMARY_CONSTRUCTOR_PARAMETER, it)!! }
-
-    private val primaryConstructorParameters: List<KtParameter>
-        get() = (declaration as? KtClass)?.primaryConstructorParameters.orEmpty()
 }
